upgrade to redux toolkit, other deps (wip)
