+++ package.json
@@ -22,10 +22,11 @@
     "@mapbox/mapbox-sdk": "^0.13.5",
     "@material-ui/core": "^1.0.0",
     "@material-ui/icons": "^1.0.0",
+    "@reduxjs/toolkit": "^1.9.0",
     "@sentry/react": "^7.14.0",
     "@sentry/webpack-plugin": "^1.19.0",
     "classnames": "^2.3.2",
-    "connected-react-router": "^4.5.0",
+    "connected-react-router": "^6.9.3",
     "debounce": "^1.2.1",
     "fecha": "^3.0.3",
     "global": "^4.4.0",
@@ -44,14 +45,12 @@
     "react-map-gl": "^5.3.15",
     "react-measure": "^2.5.2",
     "react-player": "^2.11.0",
-    "react-redux": "^5.0.7",
+    "react-redux": "^8.0.5",
     "react-responsive-carousel": "^3.2.23",
     "react-router": "^5.3.4",
     "react-router-dom": "^5.3.4",
     "react-scripts": "5.0.1",
-    "reduce-reducers": "^1.0.4",
-    "redux": "^4.2.0",
-    "redux-thunk": "^2.4.1"
+    "reduce-reducers": "^1.0.4"
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
