+++ src/store/middleware/index.js
@@ -1,0 +1,9 @@
+import { analyticsMiddleware } from './analytics';
+import { onHistoryMiddleware } from './history';
+
+const middleware = [
+  analyticsMiddleware,
+  onHistoryMiddleware,
+];
+
+export default middleware;
