+++ src/store/reducers/timeline/playback.js
@@ -1,7 +1,7 @@
 // basic helper functions for controlling playback
 // we shouldn't want to edit the raw state most of the time, helper functions are better
-import * as Types from '../actions/types';
-import store from '../store';
+import * as Types from '../../../actions/types';
+import store from '../../index';
 
 // fetch current playback offset, relative to `state.filter.start`
 export function currentOffset(state = null) {
